fix: clear flows on stop and create new instances of scope and job on start
Since AfoilComputationManager is Singleton annotated its instance will be kept
alongside the application. If user stops the current computation and starts a
new one without closing the app, the singleton instance will be kept as intended.
In this way computationJob remains cancelled even when the service restarts,
preventing further coroutine launches within the computationScope and the flows will
replay previous computation data to new subscribers.
Clearing the flow cache and creating new instances of scope and job on start avoids
the mentioned issue.